Found a better place for my tests.
